+++ test/test_show.py
@@ -15,19 +15,6 @@ import pytest
 BENCHMARK_DIR = abspath(join(dirname(__file__), 'example_results'))
 MACHINE_FILE = abspath(join(dirname(__file__), 'asv-machine.json'))
 
-@pytest.fixture
-def show_fixture(tmpdir, example_results):
-    tmpdir = str(tmpdir)
-    os.chdir(tmpdir)
-
-    conf = config.Config.from_json(
-        {'results_dir': example_results,
-         'repo': tools.generate_test_repo(tmpdir).path,
-         'project': 'asv',
-         'environment_type': "shouldn't matter what"})
-
-    return conf
-
 
 def test_show(capsys, show_fixture):
     conf = show_fixture
